RootModule should initialize PyNumpyIteratorType
